Move Resolvers to their own namespace
